Provide sparse matrix representation for W
Provide the option of using a sparse matrix representation for
the reservoir W. For low sparsity (as is typical of small-world
networks) this can provide a substantial speedup to training.

Furthermore, the loss computing frequency flag now also affects
how often the Wout matrix is computed. Typically, Wout is not
needed until the end of training, unless one wants the training
loss outputted. Reducing the times Wout is computed drastically
speeds up training.

Add the option of discarding warmup points on every batch. This
is useful if batches are not continuous in time, making it logical
to have a warmup period on every batch before learning on data.